+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -169,7 +169,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     metadata: RetrieveDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KY2MFWKQND7YDB3KHFJS77W6"
+        "01KY3S9PGBDVQC5309Q80XZTT8"
     )
     r"""The unique identifier of the data source"""