+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -183,7 +183,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     metadata: UpdateDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KY2MFWKZPDDZ506CZYCECW6P"
+        "01KY3S9PGJ75C7YHFVW6Z2F7ZM"
     )
     r"""The unique identifier of the data source"""