+++ packages/orq-rc/src/orq_ai_sdk/models/updatetoolop.py
@@ -292,7 +292,7 @@ class RequestBodyTools(BaseModel):
 
     schema_: Annotated[UpdateToolRequestBodyToolsSchema, pydantic.Field(alias="schema")]
 
-    id: Optional[str] = "01KY2MFWD4DGPSM2RBBYRJCH1H"
+    id: Optional[str] = "01KY3S9PB8W1YGHG516RHBH06F"
 
     description: Optional[str] = None
 
@@ -1362,7 +1362,7 @@ class UpdateToolResponseBodyCodeExecutionTool(BaseModel):
     code_tool: UpdateToolResponseBodyCodeTool
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KY2MFWCPM99NAR6ASGYPMDV3"
+        "tool_01KY3S9PAXEYPPG5H8BH70T4AQ"
     )
 
     display_name: Optional[str] = None
@@ -1493,7 +1493,7 @@ class UpdateToolResponseBodyTools(BaseModel):
         UpdateToolResponseBodyToolsSchema, pydantic.Field(alias="schema")
     ]
 
-    id: Optional[str] = "01KY2MFWCN1029WTN67T1YVKEC"
+    id: Optional[str] = "01KY3S9PAWENG34NATTHKTAHD5"
 
     description: Optional[str] = None
 
@@ -1637,7 +1637,7 @@ class UpdateToolResponseBodyMCPTool(BaseModel):
     mcp: UpdateToolResponseBodyMcp
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KY2MFWCJPZM8WA2XB36CVQ3Y"
+        "tool_01KY3S9PAVQ5A86XP72HQ8XN0Q"
     )
 
     display_name: Optional[str] = None
@@ -1942,7 +1942,7 @@ class UpdateToolResponseBodyHTTPTool(BaseModel):
     http: UpdateToolResponseBodyHTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KY2MFWCE9ECJDJW2EKBG2YKB"
+        "tool_01KY3S9PARP0KE149ZFV8F4DHH"
     )
 
     display_name: Optional[str] = None
@@ -2136,7 +2136,7 @@ class UpdateToolResponseBodyJSONSchemaTool(BaseModel):
     json_schema: UpdateToolResponseBodyJSONSchema
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KY2MFWCA1JZY6JACD9DJC3V9"
+        "tool_01KY3S9PAP62X38JYNQ2D2Q315"
     )
 
     display_name: Optional[str] = None
@@ -2334,7 +2334,7 @@ class UpdateToolResponseBodyFunctionTool(BaseModel):
     function: UpdateToolResponseBodyFunction
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KY2MFWC8EQSWN6HPBNZ802Q7"
+        "tool_01KY3S9PAN1MSYAPBYYF2BAB22"
     )
 
     display_name: Optional[str] = None