+++ packages/orq-rc/RELEASES.md
@@ -9609,3 +9609,13 @@ Based on:
 - [python v4.12.0-rc.49] packages/orq-rc
 ### Releases
 - [PyPI v4.12.0-rc.49] https://pypi.org/project/orq-ai-sdk/4.12.0-rc.49 - packages/orq-rc
+
+## 2026-07-22 02:10:55
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.790.3 (2.918.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.12.0-rc.50] packages/orq-rc
+### Releases
+- [PyPI v4.12.0-rc.50] https://pypi.org/project/orq-ai-sdk/4.12.0-rc.50 - packages/orq-rc